No parameters should have no coverage gap
Summary:
### Context:
In a previous task, we identified coverage gaps with `CallableParameterIsUnknownOrAny` coverage gaps. There is a bug that when a callable has no parameter, there is a `CallableParameterIsUnknownOrAny` coverage gap when this should not be the case.

### Problem:
Current code checks that every parameter is either `Any` or `Unknown` to surface this coverage gap. However, when there are no parameters, this is true for every parameter.

### Solution:
We check length to see whether there are no parameters; this should not return a coverage gap.

Reviewed By: pradeep90

Differential Revision: D37211232

fbshipit-source-id: 2a2133c5bed3802c77c9f22db77248a7089e9319