Implement query for expression level coverage
Summary:
### Context:
In previous tasks, we made functions to get coverage gaps for a module. In order to expose this functionality to the user, we want to add a query called `expression_level_coverage(paths)`.

### Problem:
In a previous diff, we set up request and response variants as well as useful types. In this diff, we need to implement the actual response.

### Solution:
In `LocationBasedLookup` we add a function to return `total_expression` and `coverage_gaps`.

In `LocationBasedLookupProcessor` we call the aforementioned function.

Finally, in `query.ml` we call the processor on each path in order to build the response containing `path,total_expression,coverage_gaps`.

### Stack diff:
In earlier diffs, we
- Add response and request variants as well as other useful types

In this diff, we
- Implement the actual response to the query

Reviewed By: pradeep90

Differential Revision: D37192286

fbshipit-source-id: c654fbf9ea593833573be7b69c460e8b3a98c438