New Implementation of Illegal Overrides
Summary:
Goal: Implement the error handling that checks if the define clause contains an override decorator illegally.

As per the code below, there are two main cases where this override decorator could be illegal.
1. None of the superclasses (and implicit superclasses, such as "object") contains the annotated method. (NothingOverridden)
2. The callable being overridden is actually a top-level function (or nested function) & not a method. (IllegalOverrideDecorator)

Note: If the method is overriding a method of a superclass that has different types, that should be handled by an existing error: Inconsistent override (error code 15)

Edit: I did include an additional error called NonMethodOverridden to include cases where a method overrode something that wasn't a method: for example, a class or a class attribute. Per comments from the team, it seems like this is something that should be implemented in InconsistentOverride instead of IllegalOverride. I wasn't sure whether to remove it or not, please let me know what you think!

Reviewed By: pradeep90

Differential Revision: D37120591

fbshipit-source-id: b185646f9fbd0d95aea4af248327b40662cd4a68