Join frames in the user-provided models with `Attach`
Summary:
This diff fixes the problem demonstrated in the last diff, via a simple but coarse approach.
In particular, within every local taint, we join every frame with the frame in the same local taint that is the `Attach` kind.
However, this fails to create `always-via` features for the situation where the `Attach` kind is specified on attributes of an argument or a return value (see function `attach_to_returned_source_ports`).
In this case, for any node in a taint tree, we want to find its parent nodes (including itself) that has a frame of the `Attach` kind, and then join these frames with the frames in the node.

Reviewed By: arthaud

Differential Revision: D37253814

fbshipit-source-id: 8c21001587a314e574de66847fc5570719161967