For every frame, convert may breadcrumbs into must breadcrumbs
Summary:
Before the analysis begins, we create a set of initial models from model generators and from `.pysa` files.
Since it is possible that a same target is specified with multiple models, we currently join all of these models.
However, it is desirable that, even if a feature `X` appears in one model, the resultant model has a feature of `always-X`, instead of `X`, which is the current behavior.
Meanwhile, this should be desirable for any initial models. That is, a user always wants an `always-X` feature in the initial models (otherwise why would he / she specify this feature?), and an `X` feature appears only in an inferred model.
This diff implements the above need, by converting the may features into must features in every frame (of a taint), after joining two models of a same target.

Reviewed By: arthaud

Differential Revision: D37226555

fbshipit-source-id: 0ec533312624f9522beb5c474a856fac56c204b8