TypeEnvironment as Environment: stop using callbacks
Summary:
This is the first commit of a stack in which I want to turn the
TypeEnvironment into a normal EnvironmentTable functor instantiation,
backed by a table of `{raw_errors, local_annotations}` records.

In this diff, I just take the existing setup of two shared-memory
tables and tweak it to make more sense by directly using the table actions
rather than making a record of callbacks to the table actions.

Reviewed By: pradeep90

Differential Revision: D37281604

fbshipit-source-id: 592ba0cfcb45bf3f045ab1061115d4087f9c21ac