TypeEnvironment as Environment: remove setters
Summary:
This is the second commit of a stack where I plan to convert
TypeEnvironment into a normal dependency tracked EnvironmentTable.

In this commit, I remove the public functions for setting errors
and local annotations; these shouldn't really be exposed, they were
an artifact from when we allowed TypeCheck to mutate the environment
from the outside.

Reviewed By: pradeep90

Differential Revision: D37281603

fbshipit-source-id: b1881deeec06d09e0515eb9951540f1f7c32a3ab