TypeEnvironment as Environment: create TypeControls.t
Summary:
The typecheck phase of environment building is fairly complex and takes
quite a few flags.

I'm trying to get away from passing the entire Configuration.Analysis.t
to code like this because that obscures what flags are actually needed;
instead, now I create a small record `EnvironmentControls.TypeCheckControls.t`
with just the information we need so that it's clear to see.

NOTE: I was considering putting `populate_defines` into this record as well but
for now I'll hold off and leave the `call_graph_builder` input so we could theoretically
use it for grey-box testing. We can come back to this later, it won't get in the way
of making TypeEnvironment overlayable so there's no need for it right now.

Reviewed By: pradeep90

Differential Revision: D37289620

fbshipit-source-id: 0f796afca9123ac3fd676e3d181ab8bfd410df5e