TypeEnvironment as environment: refactor check_define
Summary:
In order to make TypeEnvironment into a proper environment, we
want the `check_define` function to output a type that we can store
in a single table. In order to get there, it will be much easier
to construct the result just once, which we can do if we push the
try/with lower.

Reviewed By: pradeep90

Differential Revision: D37281608

fbshipit-source-id: b2f347087002a9f07bf6b2ed2b06a199c4d721d1