Make TypeEnvironment a true Environment: add EnvironmentControls
Summary:
In order to make TypeEnvironment into a true environment, we need
to make `produce_value` a pure function taking only the upstream read only.

This means that the callgraph builder, which is currently set as a flag-like argument
in `populate_for_define`, needs to be defined on the environment stack. This makes perfect
sense, but the current code structure doesn't allow it; it doesn't really make sense to put it
in `configuration` because
- it isn't serializable, and configuration needs to be
- it isn't necessarily the same for all environments in a server; for example overlays won't
  build the callgraph. But we want the configuration to be always shared.

My solution is to create another "configuration-like" type representing various levers to control environment behavior, which I name `EnvironmentControls` (I'd be okay in principle with `EnvironmentConfiguration` but I think the pun with `Configuration.Analysis.t` would be asking for confusion.)

Note that it's debatable whether this type should live *next to* configuration and contain only flags that don't belong in `Configuration.Analysis.t`, or whether it should *wrap* a configuration. I opted for the latter for a couple of reasons:
- one is that it makes the API to create an environment simpler: always pass an EnvironmentControls.t
- another is that it lets us hide details of how we use the various fields in the configuration to control environment behavior, potentially hiding destructuring and multiple-flag-reading complexity
- another is that, by making most environment-controlling behavior go through EnvironmentControls, it becomes a bit easier to see which fields in configuration control environment behavio

This commit is a pure refactor, where we wrapt the `configuration` in a layer and provide a few helper functions so that we rarely read directly from the configuration; only AstEnvironment and ModuleTracker ever need *direct* access to a configuration (that's only because they interact directly with `ModulePath.t`).

I'll add the callgraph builder in the next commit.

Reviewed By: pradeep90

Differential Revision: D37281606

fbshipit-source-id: 21fd9dfa0732e9a54aa2db1f3055da2b807676e7