TypeEnvironment as Environment: add populate_call_graph flag
Summary:
Just to recap some previous discussion leading to this diff:
- Because the call graph handling is currently done via a side effect
  where we write to a global table, it's tricky to refactor it to perform
  like a normal environment table update.
- While it would be desirable to change that, doing so isn't necessary for
  overlays and is therefore a distraction we don't want to deal with
- As a result, we want to control whether to use that side effect using
  environment-level read-only data so that when we write a
  TypeEnvironment.produce_value function it can do the right thing

With those constraints in mind, we want to use the new EnvironmentControls
module to control whether or not we build the callgraph. We do this by
adding a `populate_call_graph` flag to it, and using that in `TypeEnvironment`
to set the builder.

Note that this means in environments we can only have either no call graph
(use the NullBuilder) or yes a call graph (use the DefaultBuilder). But we
are still injecting the builder as a first-class module in `TypeCheck`, so
it's still available to use via dependency injection for things like
grey-box testing.

In addition to being simpler to reason about, a boolean flag was actually
necessary because the builder module depends on global_resolution, so trying
to inject it directly produces a dependency cycle.

Reviewed By: pradeep90

Differential Revision: D37289617

fbshipit-source-id: f652d2de0f7b1723a2e1652dc20b78e37352bbf4