TypeEnvironment as Environment: handle dependency in the right place
Summary:
For a long time now, we've been converting defines into pairs of (name,
dependency) which is very strange becuase the dependency is a *function* of the
name, so they ought to get handled near the bottom layer (and indeed when we
turn it into a dependency-tracked table, `get` will handle this for us!

In the meantime, let's do the sensible thing and create the dependency in
`populate_for_definition` by checking the `track_dependencies` field; while
we're at it let's make `~dependency` a normal named argument to
`TypeCheck.check_define_by_name`.

The calling code gets cleaner :)

Reviewed By: pradeep90

Differential Revision: D37298522

fbshipit-source-id: 7127afa263959bb3dbfc41b77c683919f4b9503d