TypeEnvironment as Environment: kill configuration argument
Summary:
All environment layers are supposed to be getting data they need out
of configuration through the environment controls (which is what
allows them to be made out of pure functions depending on upstream read-only!).

We can and should follow this pattern in TypeEnvironment.

Note that a few bad tests had to be patched because they were depending
on being able to inject a configuration *different* from the one powering
the environment, which is not a legal thing to do and shouldn't be permitted.
In particular, the parser error tests were relying on `debug` being false in
AstEnvironment, and yet they were setting `debug` to true for type errors; this
isn't useful or necessary but tripped me up for a while.

Reviewed By: pradeep90

Differential Revision: D37298523

fbshipit-source-id: 8429176fb434a9aad6158f23714280acfcf8bfbe