TypeEnvironment as Environment: clean up CheckResult type
Summary:
In this commit, we have `check_define` output a slightly different
type (with the errors optional and local annotations as read-only)
that is more suitable for storing in a single environment table.

This greatly simplifies the TypeEnvironment logic because it is
pushing the decisions about what data is needed into the computation,
as is needed for environment code.

Reviewed By: pradeep90

Differential Revision: D37281607

fbshipit-source-id: 24e8cb53c31559ca91660b72e99d7aebe7fa74d9