Remove controls from read/write API
Summary:
Most of the environment layers (everything after
AstEnvironment) should really only need controls either
at creation time or via read-only data.

Let's just get rid of the `controls` function in the
read-write API.

In the process, I caught a misnamed variable in
UnmanagedCache, where we were passing a controls value but
naming it `ast_environment`; that's probably because I had
a search-and-replace that failed to catch everything
in an earlier refactor.

Reviewed By: pradeep90

Differential Revision: D37313248

fbshipit-source-id: bee7659e4b26c23f4e24da8eed600bb30ad936a9