Do not unnecessarily depend on `compare`
Summary:
We currently insist that the Value type of a dependency-tracked
table should be comparable, where comparable means having the
`compare` function, which implies a linear ordering.

This is not necessary, and requires us to arbitrarily define
linear ordering in places where it doesn't really make sense -
all we *actually* need is an `equivalent` function (which doesn't
technically have to be `equal` - any equivalence class that preserves
sementics would do for example a location-invariant equal is typically
okay). The only place where we ever *used* this compare was inside
of an `Int.equal 0 (compare ...)` block.

So, let's just require an `equivalent` function. This will save us
from defining a pointless total ordering on `CheckResult` when all
we really need is `equal`.

Reviewed By: grievejia

Differential Revision: D37315903

fbshipit-source-id: 9fe76c69c7925351c4dd374da86a17230e882d20