Turn TypeEnvironment into an EnvironmentTable
Summary:
This is a minimal diff that turns TypeEnvironment into a proper
dependency-tracked environment table instance.

It is minimal in that we're not yet forcing consumers to work with
it like an EnvironmentTable; in particular:
- We aren't yet calling `update_this_and_all_previous_environments`
- We expose an `Unsafe` api so that we can
  - still depend on read/write access to the `global_environment`
  - still define the `invalidate` function
- We still expose side-effectful `populate_.*` functions

The next several commits will refactor callers until these
hacks can be backed out, but just making the EnvironmentTable
functor work is a significant chunk of effort that I think is best
done as its own commit.

Reviewed By: grievejia

Differential Revision: D37318837

fbshipit-source-id: 55253105ac1be5aaff9b2f530ad66bddc57abdad