ErrorsEnvironment: Refactor Preprocessing logic
Summary:
Now that TypeEnvironment is a dependency-tracked EnvironmentTable,
I'd like to go ahead and do something similar to package up the
logic and data for postprocessing (which also has to be turned into
an overlay, which will be much easier if it is constructed using
the same functor).

My intention is to call this new environment layer `ErrorsEnvironment`.

To prepare for that, a first step is to clean up the interfaces
in Preprocessing. This commit makes two changes:
- first (and most immediately beneficial because it cleans up all
  the caller code quite a bit!), we get rid of the `configuration`
  argument in `Postprocessing.run`; taking this argument was always a mistake
  because TypeEnvironment already owns a configuration and it's not actually
  well-defined what would happen if we pass in a different configuration.
- second, we prepare for a world with an `ErrorsEnvironment` that has
  a `produce_value` function by isolating the code that post-processes
  just one module into its own top-level function.
- this is purely cosmetic, but I push down some of the logic for grabbing
  various bits of data out of `TypeEnvironment.ReadOnly.t` to where they
  are actually used.

Other than stripping out the unnecessary and bad argument `configuration`,
the interface doesn't change at all.

Reviewed By: grievejia

Differential Revision: D37323527

fbshipit-source-id: d6a9973099840e59ec1549c2f9243f4470a39bcb