Move code around in `check.ml`
Summary:
This has been bothering me for a while, but now it's actually
in my way: we're currently grabbing data out of the environment
in a relatively high scope, only to define an `indices ()`
callback that only gets used in debug.

Let's just push all that down into debug. This makes it easier
to see the actual control flow.

Reviewed By: grievejia

Differential Revision: D37334501

fbshipit-source-id: 0ee1d8ea3099f225e968276fa964204a4726bf33