Stop exposing `define_additions`
Summary:
Inside of UnannotatedGlobalEnvironment we do need to compute
new defines because of the hand-rolled dependency tracking logic.

But we no longer need to expose this information in UpdateResult,
which is a major simplification - because we are now using a
dependency-tracked TypeEnvironment with a lazy `get_errors`, we'll
compute type check results for any new defines in `invalidated_modules`
in the postprocessing pass of an incremental update.

So, we can strip out this API, which is fantastic because it was one
of the more confusing aspects of `UpdateResult` behavior.

Reviewed By: grievejia

Differential Revision: D37322251

fbshipit-source-id: 530028eef9789fee0c6e84feef3a6a52b8e312bb