Adding Class Attribute to List of Symbols Being Found
Summary:
Per T114362484, Class Attributes are one of the symbols found by Pyls but not supported Pyre. This diff is to support class attributes.

To verify that an AST element is a class attribute, we check for two conditions (*both need to be true*):
1. Element is either an AnnAssign or an Assign
2. Element must appear **directly** inside of a class definition.

~~For condition #2, we need to keep track of the elements have been most recently visited. To do that, we use a **stack**.~~
~~We **push** an element onto the stack once we've begun visiting it.~~
~~We **pop** the element off the stack once we're done both visiting it, and its children.~~

Note: One thing that I would love to get feedback on from the experts on this team is the things I'm doing to get this module to pass pyre checks (including casts, and gigantic union of symbols)

Update: (6/21/22 at noon EST): we are no longer using a stack to keep track of the traversed visited elements, instead using an instance variable called "previous_element" to determine whether an item is a class attribute (taking it in the constructor of __SymbolCollector)

Reviewed By: grievejia

Differential Revision: D37174969

fbshipit-source-id: 8e071f828f12d548f59d771b73056c8e77f5c017