Refactor functions related to flag and directory
Summary:
### Context:
We want to add a cli subcommand for expression level coverage to `pyre coverage`.

### Problem:
We want to set up functions so that the `expression-level-coverage` subcommand can return different responses depending on the presence of the `--print-summary` flag and whether a path is given in the `--path flag` .

### Solution:
We want to use the functionality of `query` but be able to log the response or print a summary depending on the flag. So, we add similar functions to an `expression_level` specific file that logs the response or the summary depending on the `print_summary` parameter.

We also want to be able to find paths in the working directory, if the user does not input a path with the `--path-flag`. This functionality exists in `coverage.py`, so we extract it as a separate function.

### Stack diff:
In this diff and earlier diffs, we
- Set up the skeleton of the client subcommand
- Call backend in expression level subcommand
- Add a function that prints a human readable summary for `expression-level`
- Refactor to help support responses for flags

In later diffs, we
- Add a flag that prints a human readable summary for `expression-level`

Reviewed By: grievejia

Differential Revision: D37330518

fbshipit-source-id: 230dc347610a448199ac1fd6abb110e7143c42c0