Add function for expression level summary
Summary:
### Context:
We want the `pyre coverage expression-level` subcommand to support the `--print-summary` flag, that prints the percentage of lines covered, for each path and overall

### Problem:
We need a function that takes a response to an `expression_level_coverage` query and returns the percentage of lines covered, for each path and overall. Then, we can add functionality of the flag.

### Solution:
We add a function that takes in a response and returns percentage of lines covered, for each path and overall. We use the `dataclasses-json` library to parse the response and obtain the percent of covered expressions.

We handle edge cases for bad inputs and responses with no files/expressions.

### Stack diff:
In this diff and earlier diffs, we
- Set up the skeleton of the client subcommand
- Call backend in expression level subcommand
- Add a function that prints a human readable summary for `expression-level`

In later diffs, we
- Add a flag that prints a human readable summary for `expression-level`

Reviewed By: pradeep90

Differential Revision: D37329892

fbshipit-source-id: 416668f83a07fb307d21f483a82a00e15f6a8977