AstEnvironment dependencies [1] AstEnvironment
Summary:
We're almost ready to turn preprocessing into a normal,
dependency-tracked environment stage. But there's one problem
we have to address first: postprocessing depends on raw sources
to find ignore comments, but `AstEnvironment` doesn't track
triggered dependencies other than the self-dependencies introduced
by `WildcardImport`.

We need to fix that, which we'll do in two parts:
- add triggered dependencies to the AstEnvironment update result;
  we'll filter out the WildcardImport dependencies to avoid exposing
  internal details and taking a perf hit
- wire those dependencies up to UnannotatedGlobalEnvironment's
  UpdateResult

This commit is the first part.

Reviewed By: grievejia

Differential Revision: D37363969

fbshipit-source-id: 425e4bc1062762bc240a2552a47e8e2c6b76463e