Clean up Query: use read-only environment in LBLP
Summary:
When we created LocationBasedLookupProcessor, we had to pass
it a TypeEnvironment.t because we didn't have a read-only module
tracker interface.

This is not type safe, and the underlying architectural problems
have been fixed. Let's clean it up - if we can avoid passing read-write
environments around it will be much, much easier to reason about
our code!

Reviewed By: grievejia

Differential Revision: D37368841

fbshipit-source-id: 15e437056206ca76e3c5a27952de54479a3d11c9