AstEnvironment dependencies [2] UnannotatedGlobalEnvironment
Summary:
In this commit, I pass the triggered dependencies that AstEnvironment
now returns out through UnannotatedGlobalEnvironment.UpdateResult.

This will allow us to trigger postprocessing dependencies (which
depend on `get_raw_source` in order to handle error filtering and
unused ignore errors.

The code is a little messy because we currently have each environment
layer own its own filtering code, which makes me expose the internals
of AstEnvironment.UpdateResult so that UnannotatedGlobalEnvironment.Overlay
can overwrite it; this is something we could potentially fix later
by having each layer filter its own updates, but the peices aren't in place
for that yet and I'd rather expose AstEnvironment.UpdateResult than distrub
the progress of the v0 design of overlays.

Reviewed By: grievejia

Differential Revision: D37363968

fbshipit-source-id: 136d13b7cea48c1c1993e20499c4f9b6b54028ba