Simplify the `Check` interface
Summary:
This diff is in response to both my own difficulties and
a review comment from the introduction of ErrorsEnvironment,
where returning errors explicitly creates messy code.

Instead, let's just return the environment - which already
has the errors in a table - and extract them as needed.

This technically introduces a tiny bit more serialization
cost but it should be inconsequential on realistic codebases,
since it's just the errors (and a normal pyre run on a healthy
trunk will have at most a few errors anyway!)

Reviewed By: grievejia

Differential Revision: D37369372

fbshipit-source-id: aa1cfae7ff30014847ccd3bd7ce88aa64888324a