Refactor test setup for checking class attributes.
Summary:
Problem: The API is hard to understand.

It accepts a source and a list of classes, but there's no real way to tell what corresponds to what.

This diff just changes the testing code.

Next diff: Accept the exact class whose attributes you want to test. That makes it clearer.

Reviewed By: grievejia

Differential Revision: D37382724

fbshipit-source-id: a79698de4be2940530d593507defa3eb84b7c881