Make expression level query work for txt file
Summary:
### Context:
In previous tasks, we added a query to get expression level coverage that takes in paths of python files. We may want to query a long list of files, but the command line has a character limit, so we want to add support reading text files that contain a list of files.

### Problem:
When given an argument prefaced by a `@` such as `arguments.txt`, the query should read the contents of the file as a list of paths.

### Solution:
When the argument is prefaced by a `@`, we read the file lines as a list of file paths and throw an error if the file is not found. We discard lines that are simply whitespace. We propagate results down the pipeline by using the `Result` type to differentiate errors from non-errors.

### Stack diff:
In this diff, we
- Add query support for txt files

Reviewed By: pradeep90

Differential Revision: D37380048

fbshipit-source-id: 8f770b7b212131d64c0736d5b837ac5e1616632a