Update testing interfaces to be errors-environment-based
Summary:
We want the scratch project interactions to be managed so that
updates are always consistent, which means they should be driven
by accessing either an errors environment update result or a
read only errors environment and then using the `ErrorsEnvironment.Testing`
api to reach down to the layer we care about.

The actual test contents are unchanged, it's only the interfaces we
go through that are updated here; this unblocks using ScratchProject
to test errors environment code in the future if we want to.

Reviewed By: grievejia

Differential Revision: D37430015

fbshipit-source-id: 3a3db42a74a500d038fd28a46a22fdbe6235bd80