Add a test of ErrorsEnvironment
Summary:
We had no unit tests of the end-to-end flow including postprocessing
prior to this commit; as a result when I ripped apart the Postprocessing
API and moved most of the logic into ErrorsEnvironment I was relying
almost exclusively on the integration tests.

That's mostly okay, but I do think we ought to have some very minimal
unit tests of the end-to-end flow; it would let us catch problems
a bit faster and writing just 1 or 2 tests won't add much overhead.

Starting with a test verifying that ErrorsEnvironment triggers
postprocessing as expected in an end-to-end cold start (using the same
logic as in Check) does what we expect.

Reviewed By: grievejia

Differential Revision: D37430018

fbshipit-source-id: b3f668d32208afbf9bd395de3c21288bc29e6de3