Expand location of synthesized for-loop assignment statement.
Summary:
Problem: We were using just the location of the target variable. This made go-to-def fail for the iterator position, since that position was not covered within the statement location range.

Note: There is a slight regression in the `types(...)` query. This is because `xs` and `xs.__iter__().__next__()` share the same location, and the `types` visitor currently picks the latter. Filed T124426942 to fix the deeper problem there.

Reviewed By: grievejia

Differential Revision: D37428315

fbshipit-source-id: aa56690cedad2f9e56402678f7a4331db65a4e24