Test for the location of the for-loop preamble.
Summary:
Problem: When creating a CFG from a for-loop statement, we call `For.preamble`. That, however, drops the location of the iterator and just uses the location of the target variable. This hurts go-to-def!

This diff: Test for the location too.

Next diff: Fix the location.

Later: Test the other block preamble locations too.

Reviewed By: grievejia

Differential Revision: D37428314

fbshipit-source-id: 3adbcdddcb4f490979e99a55067775564e250cf4