Refactor read access visitor to use collection visitors
Summary:
cc grievejia on AST visitor refactoring

This diff pulls the access visitor logic into a common library so we can reuse this for find all references.

The main implementation change is that we are no longer re-doing a custom statement visitor - this can be deleted and thus avoids having to re-implement things like the match statement and we can get this for free from the visitor.

Instead, we can avoid counting assignment targets as unuinitialized by simply removing binding locations before we throw uninitialized errors.

Reviewed By: grievejia

Differential Revision: D37025500

fbshipit-source-id: 722aa8299bc6f97c9abc1c14a8f645a2194ccae6