Treat All Assignments as Symbol Type "Variable"
Summary:
I discovered that in Pyls, all of the following symbol types are all treated as `SymbolKind.Variable`
* Class Attributes
* Global Variables
* Enums
* TypeVars

Thus, one way to greatly simplify our logic in find_symbols.py is to treat all lhs of assignments to be symbols (of variable type). Previously, there was elaborate logic to check whether the immediate parent of a symbol was a ClassDef; with this new logic, that check is no longer necessary.

One side effect of this is that now all assignments, even locally scoped variable assignments will also be treated as variables (which doesn't happen in Pyls). We deemed this to be ok because Pyright also treats them as variables (thanks for checking this, Pradeep!)

Note: I will delete the following diffs because they are now obsolete: D37278648 & D37326139

Reviewed By: pradeep90

Differential Revision: D37416545

fbshipit-source-id: 3df7876fe05381144035ce26b6d409630790f386