Refactor parsing txt
Summary:
### Context:
In a previous diff, the expression level query can accept a text file when prefaced with `@`, such as in `arguments.txt.` The logic for parsing this needs to be refactored.

### Problem:
The code currently contains problematic and error prone String methods, such as `get` and correctly indexing with `sub`.

### Solution:
We instead match on `chop_prefix ` of `@`, which doesn't require propagating changes to other helper functions.

### Stack diff
- (Previous diff) Test multiple files in txt
- (This diff) Refactor code for parsing
- Refactor code for testing

Reviewed By: stroxler

Differential Revision: D37416805

fbshipit-source-id: 6071d0ad96ca154225bcf48d23665ad0ebced540