Test multiple files in txt
Summary:
### Context:
In a previous diff, we allowed the expression level query to accept an arguments of the form `arguments.txt` to read in a text file containing a list of files. To better this, we want to add tests and refactor code for parsing txt.

### Problem:
In this diff we want to add a test for text files with multiple files

### Solution:
We add a test for text files with multiple files

### Stack diff
- (This diff) test multiple files in txt
- Refactor code for parsing
- Refactor code for testing

Reviewed By: stroxler

Differential Revision: D37393369

fbshipit-source-id: 495d7848246e6be1f229e103904b4d1f407fc830