Improve DSL Testability with pyre query
Summary:
Add a new query specification to the `pyre query` command.

Running `pyre query "model_query(path, model_query_name)"` will return in JSON a list of all models generated from the query with the name `model_query_name` in the file from `path`.

`path` must be the absolute path of a directory with a valid `taint.config` file. `model_query_name` should be the name of a model query in a `.pysa` file in the directory. To find the valid `path`s to pass into `model-query`, run `pyre`, then `pyre query "validate_taint_models()"` in the directory you want to test in.

Reviewed By: arthaud

Differential Revision: D37128862

fbshipit-source-id: 4ccc01265e8d4d7815c4fab40c3183124d5a4f09