Propagate the pyre error code in model generators
Summary:
For pysa's analysis, we run a set of python model generators. Those are scripts that run in the environment of the analyzed code and generate a set of models from it. Some of those scripts use `pyre query` to request additional information about the codebase.

This diff defines a set of exit codes for these model generators, so that we can easily classify errors in production.
It also propagates a few common error codes from the `pyre incremental` command (ran before using `pyre query`).

Reviewed By: alexkassil

Differential Revision: D37476268

fbshipit-source-id: 209a790ea1c943c75c4a0a918b683f18829453e7