Make special forms point to the module where they are defined.
Summary:
Problem: For special forms, such as `Callable`, we define them specially in `missingFromStubs.ml`, with module as `typing` and location as `any`.

However, that means VS Code doesn't go to their definition.

Solution: Just translate the `any` location to the start of the file, so that the user at least goes to the `typing` file.

Reviewed By: grievejia

Differential Revision: D37459484

fbshipit-source-id: 35d2b034ea66ad63e207aba62ec3721334c79eee