Add function to resolve type for one symbol.
Summary:
Problem: We want to get hover information for just the symbol that the user has hovered over. We are currently getting all types in the module, which can keep the server busy (even though the user gets hover results from a client-side cache).

Now that we have proved a fast end-to-end, per-location flow with go-to-def, we can reuse the same for hover.

## Stack plan

+ This diff: Add a function to resolve type for a symbol.
+ Add request and response variant for a new per-location `hover_data_for_position` query. (Remember, the plan is to use beta query APIs till code-nav works end-to-end. We can consolidate and refactor them afterwards.)
+ Implement the above query.
+ Switch the client to use the above per-location query.
+ Delet the old cache-based hover code.

-------

## This diff

Note: Go-to-def can sometimes seem slow. That is not the go-to-def query being slow. It is because, in the background, the client is sending a `types(...)` query (and running `typeCoverage` locally). These two things make go-to-def requests wait. If you look at the logs, the go-to-def round-trip latency is ~20 ms even for 1k LOC files.

After migrating to per-location hover queries (this stack), we will remove much of the above latency. After disabling the *slow* libcst-based `typeCoverage` queries, we will remove all of the extra latency.

-------

Note that the test inputs are duplicated across tests for `resolve_definition` and `resolve_type`. I can DRY them out in the future. It got a bit messy when trying to merge these two.

Reviewed By: grievejia

Differential Revision: D37459487

fbshipit-source-id: 49dfc53cc95895b929894f17eaae4aeaaca25cc1