Return json error if extension not supported
Summary:
### Context:
We want to handle edge cases for the expression-level command. We examine one edge case in this diff and will examine others in later diffs

### Problem:
One edge case we examine is when the file is not of the `.py` or other supported extension.

### Solution:
We catch this scenario and throw an error.

Reviewed By: pradeep90

Differential Revision: D37418869

fbshipit-source-id: f54573981bb33a9eabae79b7d4046a9b779e4d54