Return errors as json not exception
Summary:
### Context:
One edge case we want to handle for the expression level coverage subcommand is that the response returns an error when even a single path errors, but we still want to return responses of the other error-free paths.

### Problem:
We want to return a combination of coverage information and errors, as currently we return an error if even a single path errors.

### Solution:
We make the coverage response be a list of either coverage information or errors.

Reviewed By: pradeep90

Differential Revision: D37458135

fbshipit-source-id: c034dde14b4e92f168b434a5a268822059952992