Test strict mode and no annotation
Summary:
### Context:
One edge case we want to handle for the `expression_level_coverage` query is unannotated functions should have a coverage gap even when strict mode is on.

### Problem:
We want to make sure that when there are no annotations in the signature and when strict mode is on that there should be a coverage gap, as the parameters would be of the type `Any`.

### Solution:
We add a test case ensuring that when strict mode is on and there are no annotations, this would result in a coverage gap.

Reviewed By: pradeep90

Differential Revision: D37466029

fbshipit-source-id: d12a0fbb3a50eea1bafc391ca895ea553a7e9338