Fix txt parsing bug in coverage subcommand
Summary:
### Context:
There was a bug where `arguments.txt` works for `pyre query “expression_level_coverage(‘arguments.txt’)”` but not `pyre coverage expression-level --path arguments.txt`.

### Problem:
Currently, there is a bug in the `expression-level` subcommand due to building up absolute paths regardless of whether the path given is a `textfile.txt` or a `python.py`. This is because when no path is given, we want to get the absolute paths of all the files in the current working directory.

### Solution:
We strip the `@` and add it back on after getting the path.

Reviewed By: pradeep90

Differential Revision: D37433657

fbshipit-source-id: f7c2aba85bba3dab6d38ef05a36f6fe3c8644dc0