Test backend query exception
Summary:
### Context:
One edge case we want to handle for the `expression_level_coverage` query is that when the backend query raises an exception we should return a response instead of crashing.

### Problem:
We want to make sure that backend exceptions return a response instead of crashing.

### Solution:
We add a test case ensuring this by mocking the `query_server` call.

Reviewed By: pradeep90

Differential Revision: D37475314

fbshipit-source-id: 798bc6bad79cc0b47f11c6a3589b443bbe950251