Callable coverage gap if any one parameter has gap
Summary:
### Context:
Currently, we say that there is a coverage gap if all parameters are of the type `Any/Top`.

### Problem:
If any one parameter is `Any/Top`, there may be a coverage gap as it allows for anything to be passed in, even if it is not of the correct type. However, there is not necessarily a coverage gap if there is a default value and the default value is used. Hence, there are tradeoffs for whether we should classify there to be a coverage gap if any one parameter has a gap.

### Solution:
We check that any one parameter is `Any/Top` instead checking that all parameters are `Any/Top`. This will capture most cases, as it is more rare that there is both a default value and the default value is used, and we can address default values later.

Reviewed By: pradeep90

Differential Revision: D37378767

fbshipit-source-id: 80cf212d29cfe45966d25b9ddb971f880797e3ac